change some config and use TongJi yaw_fixer

+++ src/hik_camera/src/hik_camera_node.cpp
@@ -110,9 +110,9 @@ private:
     MV_CC_GetFloatValue(camera_handle_, "ExposureTime", &f_value);
     param_desc.integer_range[0].from_value = f_value.fMin;
     param_desc.integer_range[0].to_value = f_value.fMax;
-    int exposure_time = this->declare_parameter("exposure_time", 5000, param_desc);
+    double exposure_time = this->declare_parameter("exposure_time", 5000, param_desc);
     MV_CC_SetFloatValue(camera_handle_, "ExposureTime", exposure_time);
-    RCLCPP_INFO(this->get_logger(), "Exposure time: %d", exposure_time);
+    RCLCPP_INFO(this->get_logger(), "Exposure time: %f", exposure_time);
 
     // Gain
     param_desc.description = "Gain";
